solution/scraper/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
black==24.8.0
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.112.2
h11==0.14.0
idna==3.8
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pydantic==2.8.2
pydantic_core==2.20.1
pyodbc==5.1.0
python-dotenv==1.0.1
sniffio==1.3.1
starlette==0.38.4
tomli==2.0.1
typing_extensions==4.12.2
uvicorn==0.30.6
requests==2.31.0
PyJWT==2.6.0
SQLAlchemy==1.4.46
psycopg2-binary==2.9.6
alembic==1.10.2
celery==5.2.7
redis==4.5.4
python-multipart